replication/requirements.txt
grpcio==1.51.1
protobuf==4.21.12
uuid==1.30
